Fix np.bool AttributeError (#4)
In NumPy 1.20, np.bool was deprecated in favor of the built-in bool and the explicit NumPy scalar type np.bool_.

In NumPy 2.0, np.bool was reintroduced as an alias for np.bool_, primarily for backward compatibility. So now, np.bool and np.bool_ refer to the same type.

Since the tests were running with the latest version of Numpy, this issue was not detected.